+++ src/Module.php
@@ -5,12 +5,10 @@ namespace justcoded\yii2\rbac;
 
 class Module extends \yii\base\Module
 {
-
 	public $defaultRoute = 'permissions/index';
 
 	public function init()
 	{
 		parent::init();
 	}
-
 }